moved JitInfo in ast module
