+++ src/baseline/ast.rs
@@ -1,5 +1,6 @@
 pub use self::codegen::AstCodeGen;
+pub use self::info::{generate as generate_info, JitInfo};
 
 mod codegen;
 mod expr;
-pub mod info;
+mod info;
