+++ src/baseline/ast/codegen.rs
@@ -5,12 +5,12 @@ use dora_parser::lexer::position::Position;
 
 use crate::baseline::asm::BaselineAssembler;
 use crate::baseline::ast::expr::*;
+use crate::baseline::ast::info::JitInfo;
 use crate::baseline::codegen::{
     create_gcpoint, register_for_mode, should_emit_debug, CodeGen, CondCode, ExprStore, Scopes,
     TempOffsets,
 };
 use crate::baseline::fct::{CatchType, Comment, JitBaselineFct, JitDescriptor};
-use crate::baseline::info::JitInfo;
 use crate::class::{ClassDef, TypeParams};
 use crate::cpu::{Mem, FREG_PARAMS, FREG_RESULT, REG_PARAMS, REG_RESULT};
 use crate::ctxt::VM;
